Move work-pool-workers API logic to work-pools directory
- Consolidated all work pool related API logic in single directory
- Added WorkPoolWorker interface and buildListWorkPoolWorkersQuery to work-pools.ts
- Updated queryKeyFactory to include workers and workersList methods
- Moved worker tests to work-pools.test.ts with proper integration
- Updated import in mocks/create-fake-work-pool-worker.ts
- Removed redundant work-pool-workers directory
- All tests passing, maintaining existing functionality
