--- ui-v2/src/api/work-pool-workers/index.ts
@@ -1,5 +1,0 @@
-export {
-	buildListWorkPoolWorkersQuery,
-	type WorkPoolWorker,
-	workPoolWorkersQueryKeyFactory,
-} from "./work-pool-workers";
